add back ability to know if a class is all day by using allday cert
