+++ client/src/app.tsx
@@ -52,12 +52,12 @@ const AfterSchoolProgramInvoicing = lazy(() =>
     }))
 )
 const HolidayProgramSelectionPage = lazy(() =>
-    import('./components/HolidayPrograms/holiday-program-class-selection-page.js').then((module) => ({
+    import('./components/holiday-programs/holiday-program-class-selection-page.js').then((module) => ({
         default: module.HolidayProgramSelectionPage,
     }))
 )
 const ClassDetailsPage = lazy(() =>
-    import('./components/HolidayPrograms/ClassDetails/index.js').then((module) => ({
+    import('./components/holiday-programs/class-details/index.js').then((module) => ({
         default: module.ClassDetailsPage,
     }))
 )
@@ -81,7 +81,7 @@ const SelectedProgramProvider = lazy(() =>
     }))
 )
 const CustomerBookingScreen = lazy(() =>
-    import('./components/HolidayPrograms/CustomerBookingScreen/pages/customer-booking-page.js').then((module) => ({
+    import('./components/holiday-programs/customer-booking-screen/pages/customer-booking-page.js').then((module) => ({
         default: module.CustomerBookingPage,
     }))
 )
