+++ client/src/components/holiday-programs/customer-booking-screen/components/step-3/discount-input.tsx
@@ -1,11 +1,11 @@
 import { Button, Input, Popover, Typography } from 'antd'
+import type { DiscountCode } from 'fizz-kidz'
 import { useState } from 'react'
 
 import { InfoCircleOutlined } from '@ant-design/icons'
 import { trpc } from '@utils/trpc'
 
 import { useCart } from '../../state/cart-store'
-import { calculateDiscountedAmount } from '../../utilities'
 
 const DiscountInput = ({ numberOfKids }: { numberOfKids: number }) => {
     const total = useCart((store) => store.total)
@@ -17,7 +17,16 @@ const DiscountInput = ({ numberOfKids }: { numberOfKids: number }) => {
 
     const checkDiscountCodeMutation = trpc.holidayPrograms.checkDiscountCode.useMutation()
 
-    const validateDiscount = async () => {
+    function calculateDiscountedAmount(total: number, discount: DiscountCode) {
+        switch (discount.discountType) {
+            case 'percentage':
+                return total * (discount.discountAmount / 100)
+            case 'price':
+                return discount.discountAmount
+        }
+    }
+
+    async function validateDiscount() {
         // do not allow the 'allday' discount code
         if (value === 'allday') {
             setLoading(false)
